requirements.txt
pdfplumber>=0.11.0
rich>=13.0.0
pyyaml>=6.0
